pyproject.toml
[project]
name = "ai-real-estate-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.73.1",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "openai>=2.31.0",
    "pandas>=3.0.2",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "uvicorn>=0.44.0",
]
